+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.16.1] - 2021-11-08
+
 ### Changed
 
 - When using a volume source (snapshot or other volume), LINSTOR CSI now correctly resizes the volume to the new
@@ -322,7 +324,8 @@ on storage with the following (case sensitive!) options:<!-- Needs Docs -->
 - all storageClass parameters options are now case insenstive
 - non-debug logging is less verbose in general
 
-[Unreleased]: https://github.com/piraeusdatastore/linstor-csi/compare/v0.16.0...HEAD
+[Unreleased]: https://github.com/piraeusdatastore/linstor-csi/compare/v0.16.1...HEAD
+[0.16.1]: https://github.com/piraeusdatastore/linstor-csi/compare/v0.16.0...v0.16.1
 [0.16.0]: https://github.com/piraeusdatastore/linstor-csi/compare/v0.15.1...v0.16.0
 [0.15.1]: https://github.com/piraeusdatastore/linstor-csi/compare/v0.15.0...v0.15.1
 [0.15.0]: https://github.com/piraeusdatastore/linstor-csi/compare/v0.14.1...v0.15.0